+++ R/find_all_subs.R
@@ -11,13 +11,15 @@
 #' for the search. It is possible to use "all" relationships. The default type is "is_a". A list of
 #' possible relationships can be found
 #' \href{https://docs.google.com/document/d/1_w-DwBdCCOh1gMeeP6yqGzcnkpbHYOa3AGSODe5epcg/edit#heading=h.hnsqoqu978s5}{here}.
+#' @param exclude_parent_id a logical value that specifies if the parent ID should be included in 
+#' the returned list.
 #'
 #' @return A list of double vectors containing the provided ID and all of its sub IDs. It
 #' contains one element per input ID.
 #' @importFrom dplyr select filter
 #' @importFrom magrittr %>%
 #' @importFrom purrr map
-find_all_subs <- function(data, ids, types = "is_a") {
+find_all_subs <- function(data, ids, types = "is_a", exclude_parent_id = FALSE) {
   if (!requireNamespace("igraph", quietly = TRUE)) {
     message("Package \"igraph\" is needed for this function to work. Please install it.", call. = FALSE)
     return(invisible(NULL))
@@ -38,6 +40,10 @@ find_all_subs <- function(data, ids, types = "is_a") {
       return(NULL)
     }
     r <- igraph::subcomponent(g, match(x, igraph::V(g)$name), "out")$name
+    if(exclude_parent_id){
+      r <- r[r != x]
+    }
+    
     as.numeric(r)
   })
   result